Move clean.js to tasks folder
